fix(navigation): resolve production hash targets by id
