+++ packages/farm/src/client/production-runtime.ts
@@ -4,3 +4,4 @@ export { createClientPluginManager } from "./plugin";
 export { searchParamsToObject } from "../search-params";
 export { setFarmTrailingSlashPreference } from "../trailing-slash";
 export { setFarmBasePath, stripFarmBasePath } from "../base-path";
+export { getHashTargetElement } from "./hash-target";
